Switch to railgun-js, allow callbacks, formatting.
Railgun-js is an NPM module, so its source is not baked into this one.
It's much more powerful too and allows us to further the functionality
of this module over time. The only breaking issue with it is that it
requires a domain parameter as well as the API key, but seeing as you
can still get the test domain from the web interface, this shouldn't be
an issue. There were also no callbacks being run, so I couldn't get
Mocha tests to complete- the parameter was there, just never called.